requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9

# Config & Validation
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
email-validator>=2.1.0

# Firebase & Security
firebase-admin>=6.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Web Scraping & AI
requests>=2.32.3
beautifulsoup4>=4.12.3
youtube-transcript-api>=0.6.2
yt-dlp>=2024.5.27
google-generativeai>=0.5.4

# Workflow
langgraph>=0.0.65
